Alters the examples to better handle timezones
This commit alters the page to better handle datetimes with timezones. The result will store the datetime with the local date as datetime-value and the timezone as string-value. That way the database-specific ways of doing datetime arithmentics with timezones (see examples for [MySQL](https://andreas.heigl.org/2016/04/18/timezones-and-mysql/) and [Postgres](https://andreas.heigl.org/2016/05/29/timezones-and-postgresql/)) are still possible, the datetime will not be prone to errors due to timezone-changes between creation and the actual datetime and still the application will be able to use a generic datetime-object that can then be converted to a local timezone of the current user.